# impl/src/main/resources/api-spec.yaml
openapi: '3.0.3'
info:
  version: '0.1.0'
  title: Gamificator API
  description: A gamification engine API
security:
  - bearerAuth: []
  - apiKeyAuth: []
  - apiKeySecret: []
paths:
...
  /badges:
    post:
      description: create a badge
      operationId: createBadge
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/BadgeDTO'
      responses:
        '201':
          description: created
    get:
      description: get the list of all badge
      operationId: getAllbadges
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BadgeDTO'
  /badges/{id}:
    get:
      description: get one badge, by id
      operationId: getBadge
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                $ref: '#/components/schemas/BadgeDTO'
    put:
      description: Update a badge
      operationId: updateBadge
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/BadgeDTO'
      responses:
        '200':
          description: updated
...
  /rules:
    post:
      description: create a rule
      operationId: createRule
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/RuleCreateCommand'
      responses:
        '201':
          description: created
    get:
      description: get the list of all rule
      operationId: getAllRules
      responses:
        '200':
          description: success
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RuleDTO'
  /rules/{id}:
    parameters:
      - in: path
        name: id
        schema:
          type: integer
          format: int64
        required: true
    put:
      description: update a rule
      operationId: updateRule
      parameters:
        - in: path
          name: id
          schema:
            type: integer
            format: int64
          required: true
      requestBody:
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/RuleUpdateCommand'
      responses:
        '201':
          description: created
    delete:
      description: Delete a rule
      operationId: deleteRule
      responses:
        '204':
          description: deleted
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-KEY
    apiKeySecret:
      type: apiKey
      in: header
      name: X-API-SECRET
  schemas:
...
    BadgeDTO:
      type: object
      properties:
        name:
          type: string
        imageUrl:
          type: string
          format: uri
...
    RuleCreateCommand:
      type: object
      properties:
        applicationId:
          type: integer
          format: int64
        condition:
          $ref: '#/components/schemas/ConditionDTO'
        then:
          $ref: '#/components/schemas/ActionDTO'
    RuleUpdateCommand:
      type: object
      properties:
        condition:
          $ref: '#/components/schemas/ConditionDTO'
        then:
          $ref: '#/components/schemas/ActionDTO'
    RuleDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        condition:
          $ref: '#/components/schemas/ConditionDTO'
        then:
          $ref: '#/components/schemas/ActionDTO'
...
    ActionDTO:
      type: object
      properties:
        awardBadges:
          type: array
          items:
            type: string
        awardPoints:
          type: array
          items:
            $ref: '#/components/schemas/AwardPointDTO'
    AwardPointDTO:
      type: object
      properties:
        pointScaleName:
          type: string
        value:
          type: integer
    ConditionDTO:
      type: object
      properties:
        type:
          type: string
...